Add hot and warm shapeless benchmarks
Aside from cold benchmarks, shapeless has not warm and hot benchmarks,
done similarly to
https://github.com/jvican/compiler-benchmark/blob/master/compilation/src/main/scala/dotty/tools/dotc/DottyBenchmark.scala,
which is inspired by Jason Zaugg.

Benchmarks are now declared in independent file. This help us to
modularize benchmarks better and don't add them all in the same source
file.
